Move passwordHelpers to utils
